refactor: utils folder in retrieval
