+++ cognee/modules/retrieval/__init__.py
@@ -1,1 +1,1 @@
-from .code_graph_retrieval import code_graph_retrieval
+from cognee.modules.retrieval.utils.code_graph_retrieval import code_graph_retrieval
