+++ evals/qa_context_provider_utils.py
@@ -1,7 +1,7 @@
 import cognee
 from cognee.modules.search.types import SearchType
 from cognee.infrastructure.databases.vector import get_vector_engine
-from cognee.modules.retrieval.brute_force_triplet_search import brute_force_triplet_search
+from cognee.modules.retrieval.utils.brute_force_triplet_search import brute_force_triplet_search
 from cognee.tasks.completion.graph_query_completion import retrieved_edges_to_string
 from functools import partial
 from cognee.api.v1.cognify.cognify_v2 import get_default_tasks
